+++ common/mod.ts
@@ -1,0 +1,1 @@
+export { regex } from "./regex.ts";
