+++ examples/common.ts
@@ -7,6 +7,7 @@ import { many } from "../combinators/iteration/many.ts";
 import { repeat } from "../combinators/iteration/repeat.ts";
 import { sepBy } from "../combinators/iteration/sepBy.ts";
 import { result } from "../primitives/mod.ts";
+import { regex } from "../common/regex.ts";
 
 /**
  * Represents a predicate function
@@ -35,41 +36,6 @@ type Predicate = (input: string) => boolean;
  */
 export function regexPredicate(regex: RegExp): Predicate {
   return (input: string) => regex.test(input);
-}
-
-/**
- * Creates a parser matching against a given regex
- *
- * @example
- *
- * ```ts
- * const even = regex(/^[02468]/).error("Expected an even number");
- * const { results } = even.parseOrThrow("24"); // [{value: '2', remaining: '4', ...}]
- * const { message } = even.parse("13"); // "Expected an even number"
- * ```
- */
-export function regex(re: RegExp): Parser<string> {
-  return createParser((input, position) => {
-    const match = input.match(re);
-
-    if (!match) {
-      return {
-        success: false,
-        message: `Expected to match against regex ${re}`,
-        position,
-      };
-    }
-
-    const value = match[0];
-    const newPosition = updatePosition(position, value);
-
-    return {
-      success: true,
-      results: [
-        { value, remaining: input.slice(value.length), position: newPosition },
-      ],
-    };
-  });
 }
 
 /**
